# schema/backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiResponseBoolean": {
        "properties": {
          "code": {
            "type": "string"
          },
          "data": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ApiResponseTaskDetailResponse": {
        "properties": {
          "code": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/TaskDetailResponse"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "ApiResponseTaskListResponse": {
        "properties": {
          "code": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/TaskListResponse"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "SseEmitter": {
        "properties": {
          "timeout": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "TaskCreateRequest": {
        "properties": {
          "changeType": {
            "type": "string"
          },
          "info": {
            "type": "string"
          },
          "kbId": {
            "type": "string"
          },
          "pipelineContext": {
            "additionalProperties": {
              "type": "object"
            },
            "type": "object"
          },
          "projectId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "typeId": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "TaskDetailResponse": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "currentStage": {
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "taskId": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "typeId": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "userId": {
            "format": "int64",
            "type": "integer"
          },
          "viewData": {
            "additionalProperties": {
              "type": "object"
            },
            "type": "object"
          }
        },
        "type": "object"
      },
      "TaskListItemResponse": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "currentStage": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "taskId": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "typeId": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "viewData": {
            "additionalProperties": {
              "type": "object"
            },
            "type": "object"
          }
        },
        "type": "object"
      },
      "TaskListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TaskListItemResponse"
            },
            "type": "array"
          },
          "page": {
            "format": "int32",
            "type": "integer"
          },
          "size": {
            "format": "int32",
            "type": "integer"
          },
          "total": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "TaskRetryRequest": {
        "properties": {
          "kbId": {
            "type": "string"
          },
          "projectId": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "TaskStatusUpdateRequest": {
        "properties": {
          "changeType": {
            "type": "string"
          },
          "info": {
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "viewPatch": {
            "additionalProperties": {
              "type": "object"
            },
            "type": "object"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/tasks": {
      "get": {
        "operationId": "list",
        "parameters": [
          {
            "in": "query",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "kbId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "types",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "statuses",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseTaskListResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "task-controller"
        ]
      },
      "post": {
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseTaskDetailResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "task-controller"
        ]
      }
    },
    "/api/tasks/{taskId}": {
      "get": {
        "operationId": "detail_1",
        "parameters": [
          {
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseTaskDetailResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "task-controller"
        ]
      }
    },
    "/api/tasks/{taskId}/retry": {
      "post": {
        "operationId": "retryTask",
        "parameters": [
          {
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskRetryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBoolean"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "task-controller"
        ]
      }
    },
    "/api/tasks/{taskRecordId}/status": {
      "post": {
        "operationId": "updateTaskStatus",
        "parameters": [
          {
            "in": "path",
            "name": "taskRecordId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskStatusUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBoolean"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "task-controller"
        ]
      }
    },
...
    "/sse/tasks": {
      "get": {
        "operationId": "subscribe",
        "parameters": [
          {
            "in": "query",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "kbId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/SseEmitter"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "task-sse-controller"
        ]
      }
    }
  },
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://127.0.0.1:18080"
    }
  ]
}